Create basic classes to provide properly work of /api/createAccount endpoint
Create SuperEntity and extended this with entities Account and Customer with appropriate repositories. Provided DTOs to use in clients and create method to create account based on CustomerDTO. Also adding response DTOs, exceptions and handlers. Controller and service for Account microservice are ready to work with /api/createAccount endpoint. In this commit there are also SQL script, DTOmappers and application.yml