Implement retry pattern from Resilience4j
Implementing retry pattern on http level to control resource access time, in routes configuration in Gateway server to some microsevice to train a bit. And the last one retry pattern implemetation is in accounts service .yml config, so that works along with circuit breaker but is harder to implement because it's necessary to set fallback metod for each required request.